graph construido agora sao so os algoritmos

+++ src/graph.cpp
@@ -9,7 +9,7 @@ Graph::Graph(int num, bool dir) : n(num), hasDir(dir), nodes(num+1) {
 }
 
 // Add edge from source to destination with a certain weight
-void Graph::addEdge(int src, int dest, int weight) {
+void Graph::addEdge(int src, int dest, double weight) {
     if (src<1 || src>n || dest<1 || dest>n) return;
     nodes[src].adj.push_back({dest, weight});
     if (!hasDir) nodes[dest].adj.push_back({src, weight});
@@ -65,4 +65,3 @@ int Graph::prim(int r) {
 int Graph::kruskal() {
     return 0;
 }
-
